# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MECHA API",
        "description": "API untuk Sistem Manajemen Bengkel Mobil",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/ratings": {
            "get": {
                "tags": [
                    "Rating"
                ],
                "summary": "Get semua ratings",
                "operationId": "ad64604c65f061cafbb78a893548ad91",
                "responses": {
                    "200": {
                        "description": "Data ratings retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rating"
                ],
                "summary": "Tambah rating baru",
                "operationId": "e876bdaebd2a715619d9d19e8be09398",
                "responses": {
                    "201": {
                        "description": "Rating created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/ratings/{id}": {
            "get": {
                "tags": [
                    "Rating"
                ],
                "summary": "Get rating by ID",
                "operationId": "912afb9dee8f055dde51b69b3f5057d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rating retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rating"
                ],
                "summary": "Update rating",
                "operationId": "ad9719d294e6307635669e7607d83650",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rating updated successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rating"
                ],
                "summary": "Hapus rating",
                "operationId": "9274dcb1681f116bde72d0ed39089fbf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rating deleted successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/services": {
            "get": {
                "tags": [
                    "Service"
                ],
                "summary": "Get semua services",
                "operationId": "4c175c5d6594b4a5112bbe15a46e2bb4",
                "responses": {
                    "200": {
                        "description": "Data services retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Service"
                ],
                "summary": "Tambah service baru",
                "operationId": "b05b71c30267faa468172458580fb6fc",
                "responses": {
                    "201": {
                        "description": "Service created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Service"
                ],
                "summary": "Get service by ID",
                "operationId": "42775b582b563bc36a9e06ae850100bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Service"
                ],
                "summary": "Update service",
                "operationId": "38f6a6cc2f5cd3dfdf39805e7fa21996",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service updated successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Service"
                ],
                "summary": "Hapus service",
                "operationId": "c189d6ab2d037e8fb4228a535ae6a399",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service deleted successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "name": "Token",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}